requirements.txt
Django==5.2.4
djongo==1.2.31
pymongo==4.13.2
pandas==2.3.1
python-dotenv==1.1.1
